Backend/requirements.txt
flask
openai
PyMuPDF
faiss-cpu
tqdm

